Preserve line breaks in html_strip_tags
Adds a parameter to html_strip_tags allowing users to determine whether <br> and <p> should have their line breaks preserved. This addition preserves backward compatibility by making the parameter optional; if it's omitted, it reverts to the no-replacement behavior that was standard in previous versions.

This code should catch any variant of <br> and <p>. It preserves any text between "br"/"p" and the closing > just in case future versions want to do something with it.